+++ cmd/trickster/main.go
@@ -18,16 +18,10 @@
 package main
 
 import (
-	"fmt"
 	"os"
-	"os/signal"
-	"syscall"
 
 	"github.com/trickstercache/trickster/v2/pkg/appinfo"
-	"github.com/trickstercache/trickster/v2/pkg/appinfo/usage"
-	"github.com/trickstercache/trickster/v2/pkg/config"
-	"github.com/trickstercache/trickster/v2/pkg/errors"
-	"github.com/trickstercache/trickster/v2/pkg/httpserver"
+	"github.com/trickstercache/trickster/v2/pkg/server"
 )
 
 var (
@@ -41,45 +35,10 @@ const (
 )
 
 func main() {
-	err := daemon()
+	appinfo.SetAppInfo(applicationName, applicationVersion,
+		applicationBuildTime, applicationGitCommitID)
+	err := server.Start()
 	if err != nil {
 		os.Exit(1)
 	}
-}
-
-func daemon() error {
-	appinfo.SetAppInfo(applicationName, applicationVersion,
-		applicationBuildTime, applicationGitCommitID)
-
-	// Load Config
-	conf, err := config.Load(appinfo.Name, appinfo.Version, os.Args[1:])
-	if err != nil {
-		fmt.Println("\nERROR: Could not load configuration:", err.Error())
-		if conf != nil && conf.Flags != nil && conf.Flags.ValidateConfig {
-			usage.PrintUsage()
-		}
-	}
-	if conf == nil {
-		return errors.ErrInvalidOptions
-	}
-
-	// Serve with Config
-	quit := make(chan os.Signal, 1)
-	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)
-	err = httpserver.Serve(conf, nil, nil, exitFatal)
-	if err != nil {
-		return err
-	}
-
-	select {
-	case sig := <-quit:
-		fmt.Println("signal received", sig)
-		// more cases to be added
-	}
-
-	return nil
-}
-
-func exitFatal() {
-	os.Exit(1)
 }
