+++ pkg/backends/clickhouse/clickhouse_test.go
@@ -54,7 +54,7 @@ func TestClickhouseClientInterfacing(t *testing.T) {
 
 func TestNewClient(t *testing.T) {
 
-	conf, err := config.Load("trickster", "test", []string{"-provider", "clickhouse", "-origin-url", "http://1"})
+	conf, err := config.Load([]string{"-provider", "clickhouse", "-origin-url", "http://1"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
