+++ pkg/backends/influxdb/influxdb_test.go
@@ -52,7 +52,7 @@ func TestInfluxDBClientInterfacing(t *testing.T) {
 
 func TestNewClient(t *testing.T) {
 
-	conf, err := config.Load("trickster", "test", []string{"-provider", "influxdb", "-origin-url", "http://1"})
+	conf, err := config.Load([]string{"-provider", "influxdb", "-origin-url", "http://1"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
