+++ pkg/backends/influxdb/url_test.go
@@ -44,8 +44,8 @@ func TestSetExtentInfluxQL(t *testing.T) {
 
 	expected := strings.Replace(strings.Replace(expectedTokenized, "$START_TIME$", startToken, -1), "$END_TIME$", endToken, -1)
 
-	conf, err := config.Load("trickster", "test",
-		[]string{"-origin-url", "none:9090", "-provider", "influxdb", "-log-level", "debug"})
+	conf, err := config.Load([]string{"-origin-url", "none:9090", "-provider",
+		"influxdb", "-log-level", "debug"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -91,8 +91,8 @@ var testQuery = `from("test-bucket")
 `
 
 func TestSetExtentFlux(t *testing.T) {
-	conf, err := config.Load("trickster", "test",
-		[]string{"-origin-url", "none:9090", "-provider", "influxdb", "-log-level", "debug"})
+	conf, err := config.Load([]string{"-origin-url", "none:9090", "-provider",
+		"influxdb", "-log-level", "debug"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
