+++ pkg/backends/irondb/url_test.go
@@ -44,10 +44,9 @@ func TestSetExtent(t *testing.T) {
 	stFl := time.Unix(start.Unix()-(start.Unix()%300), 0)
 	etFl := time.Unix(end.Unix()-(end.Unix()%300), 0)
 	e := &timeseries.Extent{Start: start, End: end}
-	conf, err := config.Load("trickster", "test",
-		[]string{"-origin-url", "none:9090",
-			"-provider", "irondb",
-			"-log-level", "debug"})
+	conf, err := config.Load([]string{"-origin-url", "none:9090",
+		"-provider", "irondb",
+		"-log-level", "debug"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -209,10 +208,9 @@ func TestFastForwardURL(t *testing.T) {
 	now := time.Now().Unix()
 	start := now - (now % 300)
 	end := start + 300
-	conf, err := config.Load("trickster", "test",
-		[]string{"-origin-url", "none:9090",
-			"-provider", "irondb",
-			"-log-level", "debug"})
+	conf, err := config.Load([]string{"-origin-url", "none:9090",
+		"-provider", "irondb",
+		"-log-level", "debug"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
