+++ pkg/cache/registration/registration_test.go
@@ -31,7 +31,7 @@ import (
 
 func TestLoadCachesFromConfig(t *testing.T) {
 
-	conf, err := config.Load("trickster", "test",
+	conf, err := config.Load(
 		[]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
