--- pkg/httpserver/signal/signal.go
@@ -1,65 +1,0 @@
-/*
- * Copyright 2018 The Trickster Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package signal
-
-import (
-	"os"
-	"os/signal"
-	"syscall"
-
-	"github.com/trickstercache/trickster/v2/pkg/cache"
-	"github.com/trickstercache/trickster/v2/pkg/config"
-	"github.com/trickstercache/trickster/v2/pkg/observability/logging"
-	"github.com/trickstercache/trickster/v2/pkg/observability/logging/logger"
-)
-
-var hups = make(chan os.Signal, 1)
-
-func init() {
-	signal.Notify(hups, syscall.SIGHUP)
-}
-
-type ServeFunc = func(*config.Config, map[string]cache.Cache, func()) error
-
-func StartHupMonitor(conf *config.Config,
-	caches map[string]cache.Cache, f ServeFunc) {
-	if conf == nil || conf.Resources == nil || f == nil {
-		return
-	}
-	// assumes all parameters are instantiated
-	go func() {
-		for {
-			select {
-			case <-hups:
-				conf.Main.ReloaderLock.Lock()
-				if conf.IsStale() {
-					logger.Warn("configuration reload starting now",
-						logging.Pairs{"source": "sighup"})
-					err := f(conf, caches, nil)
-					if err == nil {
-						conf.Main.ReloaderLock.Unlock()
-						return // runConfig will start a new HupMonitor in place of this one
-					}
-				}
-				conf.Main.ReloaderLock.Unlock()
-				logger.Warn("configuration NOT reloaded", nil)
-			case <-conf.Resources.QuitChan:
-				return
-			}
-		}
-	}()
-}
