+++ pkg/proxy/engines/cache_chunk_test.go
@@ -37,7 +37,7 @@ import (
 
 func TestMultiPartByteRangeChunks(t *testing.T) {
 
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -69,7 +69,7 @@ func TestMultiPartByteRangeChunks(t *testing.T) {
 func TestCacheHitRangeRequestChunks(t *testing.T) {
 	logger.SetLogger(testLogger)
 	expected := "is a "
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -109,7 +109,7 @@ func TestCacheHitRangeRequestChunks(t *testing.T) {
 
 func TestCacheHitRangeRequest2Chunks(t *testing.T) {
 	logger.SetLogger(testLogger)
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -155,7 +155,7 @@ func TestCacheHitRangeRequest2Chunks(t *testing.T) {
 
 func TestCacheHitRangeRequest3Chunks(t *testing.T) {
 	logger.SetLogger(testLogger)
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -196,7 +196,7 @@ func TestCacheHitRangeRequest3Chunks(t *testing.T) {
 
 func TestPartialCacheMissRangeRequestChunks(t *testing.T) {
 	logger.SetLogger(testLogger)
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -241,7 +241,7 @@ func TestPartialCacheMissRangeRequestChunks(t *testing.T) {
 }
 
 func TestFullCacheMissRangeRequestChunks(t *testing.T) {
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -305,7 +305,7 @@ func TestRangeRequestFromClientChunks(t *testing.T) {
 	bytes, _ := io.ReadAll(resp.Body)
 
 	//--------------------------------------
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -350,7 +350,7 @@ func TestQueryCacheChunks(t *testing.T) {
 	logger.SetLogger(testLogger)
 	expected := "1234"
 
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
