+++ pkg/proxy/engines/cache_test.go
@@ -63,7 +63,7 @@ func TestInvalidContentRange(t *testing.T) {
 func TestMultiPartByteRange(t *testing.T) {
 	logger.SetLogger(testLogger)
 
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -94,7 +94,7 @@ func TestMultiPartByteRange(t *testing.T) {
 func TestCacheHitRangeRequest(t *testing.T) {
 	logger.SetLogger(testLogger)
 	expected := "is a "
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -133,7 +133,7 @@ func TestCacheHitRangeRequest(t *testing.T) {
 
 func TestCacheHitRangeRequest2(t *testing.T) {
 	logger.SetLogger(testLogger)
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -178,7 +178,7 @@ func TestCacheHitRangeRequest2(t *testing.T) {
 
 func TestCacheHitRangeRequest3(t *testing.T) {
 	logger.SetLogger(testLogger)
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -218,7 +218,7 @@ func TestCacheHitRangeRequest3(t *testing.T) {
 
 func TestPartialCacheMissRangeRequest(t *testing.T) {
 	logger.SetLogger(testLogger)
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -263,7 +263,7 @@ func TestPartialCacheMissRangeRequest(t *testing.T) {
 
 func TestFullCacheMissRangeRequest(t *testing.T) {
 	logger.SetLogger(testLogger)
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -327,7 +327,7 @@ func TestRangeRequestFromClient(t *testing.T) {
 	bytes, _ := io.ReadAll(resp.Body)
 
 	//--------------------------------------
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
@@ -369,10 +369,9 @@ func TestRangeRequestFromClient(t *testing.T) {
 
 func TestQueryCache(t *testing.T) {
 	logger.SetLogger(testLogger)
-
 	expected := "1234"
 
-	conf, err := config.Load("trickster", "test", []string{"-origin-url", "http://1", "-provider", "test"})
+	conf, err := config.Load([]string{"-origin-url", "http://1", "-provider", "test"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
