+++ pkg/proxy/engines/httpproxy_test.go
@@ -42,8 +42,7 @@ func TestDoProxy(t *testing.T) {
 	es := tu.NewTestServer(http.StatusOK, "test", nil)
 	defer es.Close()
 
-	conf, err := config.Load("trickster", "test",
-		[]string{"-origin-url", es.URL, "-provider", "test", "-log-level", "debug"})
+	conf, err := config.Load([]string{"-origin-url", es.URL, "-provider", "test", "-log-level", "debug"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -94,7 +93,7 @@ func TestProxyRequestBadGateway(t *testing.T) {
 	const badUpstream = "http://127.0.0.1:64389"
 
 	// assume nothing listens on badUpstream, so this should force the proxy to generate a 502 Bad Gateway
-	conf, err := config.Load("trickster", "test", []string{"-origin-url",
+	conf, err := config.Load([]string{"-origin-url",
 		badUpstream, "-provider", "test", "-log-level", "debug"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
@@ -136,7 +135,7 @@ func TestClockOffsetWarning(t *testing.T) {
 	}
 	s := httptest.NewServer(http.HandlerFunc(handler))
 
-	conf, err := config.Load("trickster", "test", []string{"-origin-url",
+	conf, err := config.Load([]string{"-origin-url",
 		s.URL, "-provider", "test", "-log-level", "debug"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
@@ -154,7 +153,6 @@ func TestClockOffsetWarning(t *testing.T) {
 	r = r.WithContext(tc.WithResources(r.Context(),
 		request.NewResources(o, pc, nil, nil, nil, tu.NewTestTracer())))
 	logger.SetLogger(testLogger)
-
 	if testLogger.HasWarnedOnce("clockoffset.default") {
 		t.Errorf("expected %t got %t", false, true)
 	}
@@ -178,8 +176,7 @@ func TestDoProxyWithPCF(t *testing.T) {
 	es := tu.NewTestServer(http.StatusOK, "test", nil)
 	defer es.Close()
 
-	conf, err := config.Load("trickster", "test",
-		[]string{"-origin-url", es.URL, "-provider", "test", "-log-level", "debug"})
+	conf, err := config.Load([]string{"-origin-url", es.URL, "-provider", "test", "-log-level", "debug"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -233,7 +230,7 @@ func TestProxyRequestWithPCFMultipleClients(t *testing.T) {
 	es := tu.NewTestServer(http.StatusOK, "test", nil)
 	defer es.Close()
 
-	conf, err := config.Load("trickster", "test", []string{"-origin-url",
+	conf, err := config.Load([]string{"-origin-url",
 		es.URL, "-provider", "test", "-log-level", "debug"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
@@ -285,8 +282,7 @@ func TestProxyRequestWithPCFMultipleClients(t *testing.T) {
 
 func TestPrepareFetchReaderErr(t *testing.T) {
 	logger.SetLogger(testLogger)
-	conf, err := config.Load("trickster", "test",
-		[]string{"-origin-url", "http://example.com/", "-provider", "test", "-log-level", "debug"})
+	conf, err := config.Load([]string{"-origin-url", "http://example.com/", "-provider", "test", "-log-level", "debug"})
 	if err != nil {
 		t.Errorf("Could not load configuration: %s", err.Error())
 	}
