+++ pkg/proxy/handlers/config_test.go
@@ -27,8 +27,8 @@ import (
 
 func TestConfigHandler(t *testing.T) {
 
-	conf, err := config.Load("trickster-test", "test",
-		[]string{"-origin-url", "http://1.2.3.4", "-provider", "prometheus"})
+	conf, err := config.Load([]string{"-origin-url", "http://1.2.3.4",
+		"-provider", "prometheus"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
