+++ pkg/proxy/handlers/ping_test.go
@@ -26,8 +26,8 @@ import (
 
 func TestPingHandler(t *testing.T) {
 
-	conf, err := config.Load("trickster-test", "test",
-		[]string{"-provider", "reverseproxycache", "-origin-url", "http://0/"})
+	conf, err := config.Load([]string{"-provider", "reverseproxycache",
+		"-origin-url", "http://0/"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
