+++ pkg/proxy/listener/listener_test.go
@@ -61,7 +61,7 @@ func TestListeners(t *testing.T) {
 			Certificates: make([]tls.Certificate, 1),
 		}
 		errs <- testLG.StartListener("httpListener",
-			"", 0, 20, tc, http.NewServeMux(), wg, trs, nil, 0)
+			"", 0, 20, tc, http.NewServeMux(), trs, nil, 0)
 		close(errs)
 	}()
 
@@ -78,7 +78,7 @@ func TestListeners(t *testing.T) {
 	errs2 := make(chan error, 1)
 	go func() {
 		errs2 <- testLG.StartListenerRouter("httpListener2",
-			"", 0, 20, nil, "/", http.HandlerFunc(ph.HandleLocalResponse), wg,
+			"", 0, 20, nil, "/", http.HandlerFunc(ph.HandleLocalResponse),
 			nil, nil, 0)
 		close(errs2)
 	}()
@@ -94,7 +94,7 @@ func TestListeners(t *testing.T) {
 	}
 
 	err = testLG.StartListener("testBadPort",
-		"", -31, 20, nil, http.NewServeMux(), wg, trs, nil, 0)
+		"", -31, 20, nil, http.NewServeMux(), trs, nil, 0)
 	if err == nil {
 		t.Error("expected invalid port error")
 	}
@@ -126,7 +126,7 @@ func TestListenerAccept(t *testing.T) {
 	var err error
 	go func() {
 		err = testLG.StartListener("httpListener",
-			"", 0, 20, nil, http.NewServeMux(), nil, nil, nil, 0)
+			"", 0, 20, nil, http.NewServeMux(), nil, nil, 0)
 	}()
 	time.Sleep(time.Millisecond * 500)
 	if err != nil {
@@ -184,7 +184,7 @@ func TestListenerConnectionLimitWorks(t *testing.T) {
 	es := httptest.NewServer(http.HandlerFunc(handler))
 	defer es.Close()
 
-	_, err := config.Load("trickster", "test", []string{"-origin-url", es.URL, "-provider", "prometheus"})
+	_, err := config.Load([]string{"-origin-url", es.URL, "-provider", "prometheus"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
