+++ pkg/routing/routing_test.go
@@ -67,8 +67,7 @@ func TestRegisterProxyRoutes(t *testing.T) {
 	logger.SetLogger(logging.ConsoleLogger(level.Error))
 	var proxyClients backends.Backends
 
-	conf, err := config.Load("trickster", "test",
-		[]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "prometheus"})
+	conf, err := config.Load([]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "prometheus"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -160,8 +159,7 @@ func TestRegisterProxyRoutes(t *testing.T) {
 
 func TestRegisterProxyRoutesInflux(t *testing.T) {
 	logger.SetLogger(logging.ConsoleLogger(level.Error))
-	conf, err := config.Load("trickster", "test",
-		[]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "influxdb"})
+	conf, err := config.Load([]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "influxdb"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -183,8 +181,7 @@ func TestRegisterProxyRoutesInflux(t *testing.T) {
 
 func TestRegisterProxyRoutesReverseProxy(t *testing.T) {
 	logger.SetLogger(logging.ConsoleLogger(level.Error))
-	conf, err := config.Load("trickster", "test",
-		[]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "rp"})
+	conf, err := config.Load([]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "rp"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -206,8 +203,7 @@ func TestRegisterProxyRoutesReverseProxy(t *testing.T) {
 
 func TestRegisterProxyRoutesClickHouse(t *testing.T) {
 	logger.SetLogger(logging.ConsoleLogger(level.Error))
-	conf, err := config.Load("trickster", "test",
-		[]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "clickhouse"})
+	conf, err := config.Load([]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "clickhouse"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -227,8 +223,7 @@ func TestRegisterProxyRoutesClickHouse(t *testing.T) {
 
 func TestRegisterProxyRoutesALB(t *testing.T) {
 	logger.SetLogger(logging.ConsoleLogger(level.Error))
-	conf, err := config.Load("trickster", "test",
-		[]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "alb"})
+	conf, err := config.Load([]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "alb"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -252,8 +247,7 @@ func TestRegisterProxyRoutesALB(t *testing.T) {
 
 func TestRegisterProxyRoutesIRONdb(t *testing.T) {
 	logger.SetLogger(logging.ConsoleLogger(level.Error))
-	conf, err := config.Load("trickster", "test",
-		[]string{"-origin-url", "http://example.com", "-provider", "irondb", "-log-level", "debug"})
+	conf, err := config.Load([]string{"-origin-url", "http://example.com", "-provider", "irondb", "-log-level", "debug"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -274,8 +268,7 @@ func TestRegisterProxyRoutesIRONdb(t *testing.T) {
 
 func TestRegisterProxyRoutesWithReqRewriters(t *testing.T) {
 	logger.SetLogger(logging.ConsoleLogger(level.Error))
-	conf, err := config.Load("trickster", "test",
-		[]string{"-config", "../../testdata/test.routing.req_rewriter.conf"})
+	conf, err := config.Load([]string{"-config", "../../testdata/test.routing.req_rewriter.conf"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -304,7 +297,7 @@ func TestRegisterProxyRoutesMultipleDefaults(t *testing.T) {
 	expected2 := "only one backend can be marked as default. Found both test2 and test"
 
 	a := []string{"-config", "../../testdata/test.too_many_defaults.conf"}
-	conf, err := config.Load("trickster", "test", a)
+	conf, err := config.Load(a)
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -352,7 +345,7 @@ func TestRegisterProxyRoutesInvalidCert(t *testing.T) {
 	}
 
 	a := []string{"-config", confFile}
-	conf, err := config.Load("trickster", "test", a)
+	conf, err := config.Load(a)
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -372,7 +365,7 @@ func TestRegisterProxyRoutesInvalidCert(t *testing.T) {
 func TestRegisterProxyRoutesBadCacheName(t *testing.T) {
 	expected := `invalid cache name "test2" provided in backend options "test"`
 	a := []string{"-config", "../../testdata/test.bad_cache_name.conf"}
-	_, err := config.Load("trickster", "test", a)
+	_, err := config.Load(a)
 	if err == nil {
 		t.Errorf("expected error `%s` got nothing", expected)
 	} else if err.Error() != expected {
@@ -384,7 +377,7 @@ func TestRegisterProxyRoutesBadProvider(t *testing.T) {
 	logger.SetLogger(logging.ConsoleLogger(level.Error))
 	expected := "unknown backend provider in backend options. backendName: test, backendProvider: foo"
 	a := []string{"-config", "../../testdata/test.unknown_backend_provider.conf"}
-	conf, err := config.Load("trickster", "test", a)
+	conf, err := config.Load(a)
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -403,7 +396,7 @@ func TestRegisterProxyRoutesBadProvider(t *testing.T) {
 func TestRegisterMultipleBackends(t *testing.T) {
 	logger.SetLogger(logging.ConsoleLogger(level.Error))
 	a := []string{"-config", "../../testdata/test.multiple_backends.conf"}
-	conf, err := config.Load("trickster", "test", a)
+	conf, err := config.Load(a)
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -420,7 +413,7 @@ func TestRegisterMultipleBackends(t *testing.T) {
 func TestRegisterMultipleBackendsPlusDefault(t *testing.T) {
 	logger.SetLogger(logging.ConsoleLogger(level.Error))
 	a := []string{"-config", "../../testdata/test.multiple_backends_plus_default.conf"}
-	conf, err := config.Load("trickster", "test", a)
+	conf, err := config.Load(a)
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -442,8 +435,7 @@ func TestRegisterPathRoutes(t *testing.T) {
 	p := map[string]*po.Options{"test": {}}
 	RegisterPathRoutes(nil, nil, nil, nil, nil, p, nil, "")
 
-	conf, err := config.Load("trickster", "test",
-		[]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "rpc"})
+	conf, err := config.Load([]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "rpc"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
@@ -479,8 +471,7 @@ func TestValidateRuleClients(t *testing.T) {
 	var cl = backends.Backends{"test": c}
 	rule.ValidateOptions(cl, nil)
 
-	conf, err := config.Load("trickster", "test",
-		[]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "rpc"})
+	conf, err := config.Load([]string{"-log-level", "debug", "-origin-url", "http://1", "-provider", "rpc"})
 	if err != nil {
 		t.Fatalf("Could not load configuration: %s", err.Error())
 	}
